:recycle: refactor shared code
